requirements.txt
python-dotenv
flake8
deepl
pytest